fix V2 signature calculation issue caused by excluded query params
This change fixes an issue with the SigV2 signature calculation logic
where the prefix symbol ("?" or "&") used when constructing the
canonicalized resource string was sometimes incorrect.

Preiously, the exclusion of irrelevant query paramerers was not taken
into account when determining whether to prepend a "?" (for the first
included parameter) or "&" (for the rest).

As a result, for URLs such as "?versions&max-keys=100", the "versions"
parameter would be considered to have an index of 1 since it's
lexicographically greater than "max-keys" despite "max-keys" being
excluded. This would produce a canonicalized resource string with
"&versions" rather than the expected "?versions".