properly handle X-Amz-Content-Sha256 header for V4 presigned URLs
This change ensures that the X-Amz-Content-Sha256 request header is
properly handled for presigned URLs that use the SigV4 algorithm.

Previously, the value of this header was ignored, and any request with
such a URL was assumed to have an unsigned payload. The default payload
hash "UNSIGNED-PAYLOAD" was used in place of the payload hash specified
in the request, which caused the server to calculate a different
signature than the one provided, resulting in signature mismatch
errors.

Additionally, because the payload hash specified in the request was
ignored, no checksum verification was performed on the payload.